+++ src/table_heap.rs
@@ -1,6 +1,6 @@
 use crate::buffer::BufferPoolManager;
 use crate::disk::{PageId, RecordId};
-use crate::storage::Row;
+use crate::btree::Row;
 use crate::table_page::TablePage;
 use std::sync::{Arc, Mutex};